Remove vendored proguard annotation
Summary:
There have been multiple complaints about combining RN with various
other FB libraries, including Litho, because of bundled dependencies
that can't be deduplicated by Gradle.

This is one of three current conflicts:

1) Proguard annotations (this PR)
2) Yoga
3) fbjni

While the Yoga group name doesn't make a massive amount of sense
it was the easiest existing package to use and since we don't
have a better namespace for this, we might as well use this.

A similar change to Litho is landing right now.

Test Plan:
```
yarn
./gradlew :RNTester:android:app:assembleDebug
```