Announce state changes happening in the background
[Problem]
Currently the react native framework doesn't handle the state
changes of the focused item that happen not upon double tapping. Screen
reader doesn't get notified when the state of the focused item changes
in the background.

[Solution]
On iOS, post a layout change notification for every state changes.
On Android, send a click event whenever state "checked", "selected" or
"disabled" changes. In the case that such states changes upon user's
clicking, the duplicated click event will be skipped by Talkback.

[Testing]
Add a nested checkbox example which state changes after a delay in the
AccessibilityExample.