Fix disappearing alert bug on iOS
Present each alert (UIAlertController) in a separate window so that they are
not affected by modals or other presented view controlelrs. This also allows
one to display mulitple alerts on top of each other (it is sometimes useful).

I basically copied a bunch of code from RCTDevLoadingView to chieve this.

This patch fixes issue #22237.