Fix modals not being dismissed properly
Improve the algorithm for determining the view controller for presentation of
RCTModalHostViewController.

The problem was that RCTModalHostViewManager always used the first responder
view controller ([modalHostView reactViewController]), whereas it is possible
that such controller also had presented another view controller on top of
itself (and that one could present yet another VC, and so on). So we walk
through the presentedViewController chain to find the top-most presented VC.