Enable abstract locations for 50% of users in xplat/js
Summary:
Now that types-first is out in xplat, we can turn abstract locations on for improved recheck optimizations.

This was rolled out in www before we had the rollouts feature. This is a chance to get better data on how abstract locations actually performs.

Changelog: [Internal] Update Flow options to improve recheck speed

Reviewed By: panagosg7

Differential Revision: D18145269

fbshipit-source-id: 33941251288884ded4b0d8c9fe7df8225d3429f6