Moving some workarounds and fixes from RCTMultilineTextInputView to RCTUITextView
Summary:
The class `RCTUITextView` is designed to be a home for workarounds. In Fabric we will use it as well, so we need to have all workarounds there.
That's just a move of some code between clases, it does not really affect how the workarounds work.

Changelog: [Internal] Fabric-specific internal change.

Reviewed By: sammy-SC

Differential Revision: D18607833

fbshipit-source-id: 1ba0e8c5cd928f3bf6fe1d2727325ecfc1dbf684