[RCTImageLoader] Make C++ requirement opt-in
When building as a framework these headers get automatically added to
the framework umbrella header for React-Core. Instead of converting all
the React sources to ObjC++ files and still forcing external users that
build native source (and link against a framework build) to also
compile as ObjC++, this makes the attribution related methods opt-in to
ObjC++ builds.