[RNTester] Add CocoaPods based macOS target. (#211)
* [RNTester] Check-in pod install artefacts

* [RNTester] Add CameraRoll pod dependency back

At least for now so we’re in sync with upstream v0.60.0

* [RNTester] First pass at adding CocoaPods based macOS target

* [RNTester] Add notifications pod to macOS target

* [CocoaPods] Make RCTText pod build for macOS

* [CocoaPods] make RCTLinking pod build for macOS

* [CocoaPods] Exclude more files from RCTText pod

* [CocoaPods] Make Core pod build for macOS

* [CocoaPods] Exclude turbomodule sample from macOS

* [RNTester] Update CocoaPods artefact

* [RNTester] Update macOS target to follow non-pods version

* [CocoaPods] Tag all macOS changes to ease updating

https://github.com/microsoft/react-native/issues/214

* [RNTester] Remove development team references

This change is already made upstream in commit
https://github.com/facebook/react-native/commit/9ece5bda, so when that
is merged in this commit can be skipped.

Until then, not making this change will just lead to file thrashing.

* [RNTester] Import Xcode artefact

This change is already made upstream in commit
https://github.com/facebook/react-native/commit/c1845810, so when that
is merged in this commit can be skipped.

Until then, not making this change will just lead to file thrashing.

* [RNTester] Start Metro on launch of macOS product

* [RCTRedBox] Fix -[RCTRedBoxWindow dismiss] on macOS

As per https://www.thecave.com/2015/08/10/dispatch-async-to-main-queue-doesnt-work-with-modal-window-on-mac-os-x/#update2
the main queue is already executing a block when we start the new
modal runloop and as the main queue is a serial queue it won’t perform
new work until the current block is finished. In short, the task is
queued, but will never be performed by GCD. This change side-steps GCD
and instead directly invokes -dismiss on the main thread.