Issue Template: Add Needs Triage label to new issues (#27842)
Summary:
A new triage group is being set up. As part of the triage process, all new issues will be tagged as "Needs: Triage :mag:". We can achieve this by baking the label into the default issue template.

The `react-native-bot` has been updated so it does not auto-close issues tagged with this label. Previously, it would look for the presence of the "Bug" label to determine if someone used the default issue template.

I've removed HTML comments from the template to minimize confusion. Folks that may not be familiar with HTML or Markdown and how they handle comments.

## Changelog

[Internal] - Modify GitHub issue template
Pull Request resolved: https://github.com/facebook/react-native/pull/27842

Test Plan: N/A

Differential Revision: D19629687

Pulled By: hramos

fbshipit-source-id: 25a989ebc04e42e7be234244a9ded48756cb9ba4