Add global window to jest setup
`window` exists in the React Native runtime, but not inside the test
environment. Many libraries use `typeof window === 'undefined'` to
check if code is running in SSR. Because of the difference in the real
environment and test environment, tests can behave different than the
real app, which is an unwanted behavior.