Fix CircleCI by adding dependency on JSI to cxxreact
Summary:
D21908523 added an implicit dependency on `jsi.h` to use functions like `asObject`, etc. For some reason this doesn't break the build with BUCK (??) but it does with cocoapods. Adding the dep to the cxxreact podspec and regenerating offline mirrors to unbreak CircleCI. Also adding the BUCK dep and include statement for good measure.

Changelog: [Internal]

Differential Revision: D21924592

fbshipit-source-id: 60da149f0bdef5cb07fffae3a5fe930440138c44