fix Gravity.CENTER and canScrollHorizontally issue
canScrollHorizontally(-1) detects if a TextInput can scroll to the right,
but it does not work with Gravity.CENTER. Consequence is that the TextInput can
not be scrolled right and the parent component (in our example
a ScrollView) is instead scrolled. For this reason the TextInput
Gravity is set to Gravity.LEFT when focused for editing and set back to
Gravity.CENTER once the focus leaves the TextInput.

canScrollHorizontally(1) is used for detecting that a TextInput can scroll left, but
does not work with Gravity.CENTER. The TextInput will not scroll to the
last character and canScrollHorizontally(1) returns allways true with
Gravity.CENTER. Consequence is scroll is never delegated to the parent View (ScrollView
in our example) and the parent view (ScrollView) can not be scrolled
left once the end of the TextInput is reached.

I was aware of this issue and fixed previously by with
setGravity.(Gravity.LEFT), but I did not use the correct event for
settings the gravity back to Gravity.CENTER.

This commit uses event onScrollChanged to set the gravity back to
Gravity.CENTER if previously resetted. The gravity is set back to
Gravity.CENTER once the focus leaves the TextInput and the Scrolling is
ended.