Fix loading from Metro in Bridgeless mode
Summary:
Problem Statement: A native module needs to call a function on `ReactInstance` (in this case `loadScript`). Typically, this is handled by the bridge.
Current Bridgeless Solution: Create a new protocol (in this case `RCTJSScriptLoaderModule`) which lets a block be passed in TM init to forward the method call to `ReactInstance`. This is the best thing I could think of right now.

Reviewed By: RSNara

Differential Revision: D22512748

fbshipit-source-id: c8d9f08e73c86e7a8a95be791b6cbed4f7a67cb2