API < 28 - convert numeric font values to Typeface
Typeface.create method accepting integer weight parameter was added in
API 28. For this reason is not possible to use integer values for font
weight in react-native api {fontWeight: 400}.
https://developer.android.com/reference/android/graphics/Typeface#create(android.graphics.Typeface,%20int,%20boolean)

The workaround is converting all fontWeight > 699 to fontWeight bold.
This causes a several number of issues, this commit handles all
scenarios.

The only scenario not solved in this PR is the one of using custom font
{fontStyle: "rubiko", fontWeight: 700}, which in my opinion is a new
requirement, as using numeric values is yet not supported in
react-native.

For time constraints I did not yet solve this issue which is mainly
caused by this line. Converting the weight from numeric to the correct
Typeface value should solve the issue.

https://github.com/facebook/react-native/blob/314dbe84699ce905a94ff99e4458a55557413797/ReactAndroid/src/main/java/com/facebook/react/views/text/ReactTypefaceUtils.java#L110

if (family != null) {
  typeface = ReactFontManager.getInstance().getTypeface(family, newStyle, weight, assetManager);
}