Fix race-condition when loading local images
In local images, the callback is called inline, and so requestToken is always nil at the time the block is executed. In this case, the requestToken is irrelevant because the task is completed before it could be canceled. So initializing it with a do-nothing block satisfies the delegate.

It could also be a race condition even if the block is called on another thread, and so I also add @synchronized to protect against that race as well.

Fixes #29364 "Missing request token for request: <NSURLRequest: 0x60000253e5a0> { URL: file://"