Change default PixelRatio.getFontScale scale to 1
In my testing, doing `intendedfontsize * PixelRatio.getFontScale()` is how this is intended to be used.

It doesn't make sense to me to default the font scale to the device pixel ratio. In practice, that means font sizes will be scaled different depending on pixel density, which doesn't make sense (the "scale" of fonts should be consistent independent of pixel ratio). I'm surprised this hasn't come up before, but I think it's because in practice a font scale is always set.

I'm unable to handle a "default" font scale, because there's no way to distinguish it defaulting to pixel density vs a user setting their font scale to the same value.