check for reactTextUpdate.length
JavaScript Controller TextInput keep their own state and value for the
TextInput. This means that the value in Java

"This is my word"

can be different from the value in JavaScript, in the case of the
relevant example we delete spaces from sentence like below

"Thisismyword"

The two strings have different length, this causes several problems, an
example is

setSpan(5 ... 5) ends beyond length 4
https://github.com/facebook/react-native/pull/29070#issuecomment-647716585

where we change in JavaScript the TextInput value, but then a
setSelection or other operations are triggered in Java with the old
parameters.

Ideally would be great keeping the same state between Java and
JavaScript, but the current solution is checking and handling this
scenarios