Fix modal RedBox not running main dispatch queue (#609)
The RedBox could get an app in a “stuck” state because it would completely pause the main dispatch queue. This was especially evident by changing a style to be an incompatible value (e.g. `width: '50px'`), which would get you in a state where you'd need to just kill and restart the application.

Calling `[NSApp runModalForWindow:]` will result in a new runloop that does not process the main dispatch queue. This change instead runs a modal session on the main runloop that still processes the main dispatch queue.

Further reading:
- https://www.thecave.com/2015/08/10/dispatch-async-to-main-queue-doesnt-work-with-modal-window-on-mac-os-x/
- https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/WinPanel/Concepts/UsingModalWindows.html