Breaking a retain cycle between `RCTSurfaceTouchHandler` and `RCTFabricSurface` (another approach)
Summary:
This is a different approach to break the retain cycle: instead of detaching a view manually, we store a pointer to a view weakly.

I am working on some change that will make possible (and make use of it) to call `start` and `stop` on Surface objects mutiple times, so to make it happens we need to remove detaching from `stop` method.

Changelog: [Internal] Fabric-specific internal change.

Reviewed By: sammy-SC

Differential Revision: D24041524

fbshipit-source-id: 700b13d715d2eb84d52bf3e51e0fd2896a55baf8