Improving error messages for URL.js polyfill
`URL` and `URLSearchParams` are polyfilled in `URL.js`, but many functions are "not implemented" to keep the polyfill light.

When a developer hits this issue, it's often a bit hard to track down exactly what's wrong:
* What is `not implemented`?
* My URL parsing code has several calls to `URL`, which one is blowing up?

This PR simply makes the error messages more explicit.  This helps with grepping for your error and faster debugging.