Invoke `node` directly in generate-specs.sh
Codegen script for Xcode invokes `yarn --silent node <...>`, which breaks both when Yarn is not installed (if NPM is used), or when Yarn v2 ("Berry") is active. This commit changes generate-specs.sh to use `node` set via a `NODE_BINARY` env var, or, if it is not defined, find it with `command -v node`.