When surface is stopped, reset View ID of ReactRootView<Replace this line with a title. Use 1 line only, 67 chars or less>
Summary:
If there's a JS/native crash and the surface is stopped, the ReactRootView could be reused when the surface is restarted (or potentially to show the errors in a logbox; I'm don't think that's possible, but not 100% sure).

Regardless, now it will crash in those cases because of T83802049. There's no reason to do that, so just mark the ReactRootView as explicitly unused once the surface is stopped.

Changelog: [Internal]

Differential Revision: D26469742

fbshipit-source-id: 130b72ad765de44e4729dc37d34b22ad220ec301