Make ScrollViewStickyHeader work on android
Summary:
On Android, the zIndex alone does not bring the sticky header on top of
all rows that follow inside of the scrollview. Using elevation in
conjunction with zIndex will resolve this issue so that both, ios and
android behave exactly the same.
Also removed the relative positioning as it was causing visual issues
as well. Only zIndex and elevation are needed to make both platforms
match.