Guard .asObject calls with .isObject check in JSIExecutor
Summary:
When we call JSIExecutor::nativeCallSyncHook, we assume that the third argument is an object and call Value::asObject on it, before checking if the Object is an Array. Calling Value::asObject throws an error if the Value isn't an Object.

This diff includes an isObject check on the third argument.

Changelog: [Internal]

Reviewed By: fkgozali

Differential Revision: D26735262

fbshipit-source-id: 96eb43d6c8bc1d78f3b5e0dc24ed6d419a446ecf