Avoid redundant recursion in RCTView accessibilityLabel
If `subview` is a `RCTView` or subclass, there's no need to trigger the
recursion again, because it already did implicitly when accessing the
`.accessibilityLabel` property in the previous line.

This greatly improves performances (and stability) when testing
ReactNative apps using XCTest or Appium, which heavily rely on
accessibility queries in order to interact with UI elements.