Add RCT_EXTERN_REMAP_METHOD
The macro _RCT_EXTERN_REMAP_METHOD stated that RCT_EXTERN_REMAP_METHOD exists, but it does not.
This adds in the macro that allows for externing a method but choosing a custom JavaScript name.
Our company is writing all of our modules in Swift, so this is useful so we don't have to set all
of the argument labels properly to match up the Swift signature with the ObjC one.