Set text color to white instead of dark grey
Hello! I'm working on the Messenger Desktop team at Facebook, and I noticed that the message text in the LogBox is not very visible on the dark background.
https://ibb.co/Hg0WYj9

Proposing this change so that the text is more visible- not sure this is the right change (or color), but getting my foot in the door! :)