Add 'automaticallyAdjustsScrollIndicatorInsets' prop to ScrollView (for iOS)
Summary:
In iOS 13 Apple added a new property to UIScrollView that "automatically" alters the scroll indicator insets in a fashion similar to the way 'contentInsetAdjustmentBehavior' alters content insets.  See here for iOS documentation:
https://developer.apple.com/documentation/uikit/uiscrollview/3198043-automaticallyadjustsscrollindica?language=objc

The OS default value for this property is `true`, which we preserve. When set to `false`, the behavior matches iOS <= 12.

Closes #28140