Downgrade warning about non-clippable views being clipped to no-crash exception
Summary:
This is crashing too much in debug, which is good signal but making it harder to test, and test unrelated features.

We have some good data about this internally and validated that it's useful; we can follow up on the logged soft exceptions without actually crashing now.

Changelog: [Internal]

Differential Revision: D29698447

fbshipit-source-id: 61387c18f17f76e5de60baa1fd3c94028229c0f6