Make links independently focusable by Talkback (#31757)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/31757

This follows up on D23553222 (https://github.com/facebook/react-native/commit/b352e2da8137452f66717cf1cecb2e72abd727d7), which made links functional by using Talkback's Links menu.  We don't often use this as the sole access point for links due to it being more difficult for users to navigate to, and easy for users to miss if they don't listen to the entire description, including the hint text that announces that links are available.

Instead, we generally allow links to be focusable after the main text that contains them is focused. This diff adds this functionality for both Paper and Fabric, and also retains the existing Links menu functionality as well, for users who prefer to use it.

Reviewed By: yungsters

Differential Revision: D28691177

fbshipit-source-id: 55e675cf84df5dda3b15d81a7f9bf1f071006bd2