Cleanup `ndkBuild` task invocation from build.gradle
Summary:
This diff will remove the custom `Exec` tasks that are invoking `ndk-build`
manually right now and instead rely on the AGP apis to do so.

I've added the `externalNativeBuild` block to do pass the `ndk-build` arguments and removed those that are already applied by AGP automatically (like `NDK_DEBUG` or so).

Changelog:
[Internal] - Cleanup ndkBuild command to use AGP APIs

Differential Revision: D30107012

fbshipit-source-id: 9b893c5fcb7e018150f4255e5d271344e6b9b760