Setup Gradle Wrapper inside packages/react-native-gradle-plugin/
Summary:
This diff is setting up the Gradle wrapper inside packages/react-native-gradle-plugin/.
Currently the `./gradlew` file is missing so it's not possible to `cd` inside the project and build it from
there (also opening that folder inside the IDE will trigger the creation of the files I added here).

This diff is fixing. The files have been generated with the command:

```
gw -p packages/react-native-gradle-plugin wrapper --gradle-version 7.0.2 --distribution-type=all
```

Changelog:
[Internal] [Added] - Setup Gradle Wrapper inside packages/react-native-gradle-plugin/

Differential Revision: D30730145

fbshipit-source-id: 038060c7d7a8aaf96a14a72f637633be161ab2a0