Fix broken Android CI due to broken `BundleJsAndAssetsTask` override
Summary:
This Diff fixes the `test_android` CircleCI that is currently broken due to a faulty `BundleJsAndAssetsTask`
(that's my fault sorry for this).

The CI is failing with a `execCommand == null!` error message. The message is happening as the aformentioned
task is not correctly overriding `exec()` but just declaring a `TaskAction`. I'm fixing it.

Changelog:
[Internal] [Changed] - Fix broken Android CI due to broken `BundleJsAndAssetsTask` override

Differential Revision: D30899742

fbshipit-source-id: ddfd35dea054f70596a7092b8bc9b8ba3ec2a3c5