Simplify the Circle CI setup steps (#32207)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/32207

This is just a tentative run to remove some of the unnecessary steps on our CircleCI setup.
Specifically all the `setup_` steps are just checkin out the code on different runners and nothing more
therefore I'm cleaning that up.

I'll do a couple of runs to understand if that is also impacting performances or not.

Changelog:
[Internal] [Changed] - Simplify the CircleCI setup steps

Differential Revision: D30932949

fbshipit-source-id: 5526ea8f8e185eb5d54f12bbe241376ee3e6cb7f