fix update-ruby.sh to use proper bundle config
Rubygem's bundle will only use config as described in:
https://bundler.io/man/bundle-config.1.html#DESCRIPTION

1. Local config (<project_root>/.bundle/config or $BUNDLE_APP_CONFIG/config)
2. Environmental variables (ENV)
3. Global config (~/.bundle/config)

While in the root directory it was using the #1 (.bundle/config), in
the template it wasn't since the folder is `_bundle` (underscore, not
a leading dot).

This commit makes sure the desired config is used, also takes care to
sync both configuration files, copying the root to the template.