Make Danger Phabricator-friendly.
Summary:
This Diff adapts the Dangerfile to be a bit more friendly
for PRs that are exported from Phabricator. Specifically,
I'm disabling the check for Test Plan as those are not exported
externally. The check for Summary and Changelog are instead adapted
to work with the format exported for Phabricator.

Changelog:
[Internal] [Changed] - Make Danger Phabricator-friendly

Differential Revision: D32594622

fbshipit-source-id: 08727587357bdfd01ed7ef6428db634004d19ab5