fix(iOS): remove alert's window when call to `hide`.
Resolves this issue: #32304.

Without this change, calling to hide an alert, leaves a `UIWindow`
that blocks user interactions with the screen.

The correct way to remove a `UIWindow` in iOS is to set its
hidden property to `YES`.

Also, it is required to remove all references to the window (the
associated `windowScene` for example) and ARC will automatically
free this `UIWindow`.

Co-authored-by: paddlefish <paddlefish@users.noreply.github.com>