Allow to setup a Gradle Enterprise Instance via env variable.
Summary:
This diff allows to setup a Gradle Enterprise instance
either locally or on CI via the `GRADLE_ENTERPRISE_URL`.

Moreover, it applies the `com.gradle.enterprise` Gradle
plugin by default. This has the positive side effect of
not invalidating build cache for the included build if
you happen to pass the `--scan` flag in OSS (as the
classpath isn't changed).

Changelog:
[Android] [Added] - Allow to setup a Gradle Enterprise Instance via env variable.

Reviewed By: mdvacca

Differential Revision: D33582388

fbshipit-source-id: 0b4bf5748c7974f6354b3553818c03ef2764be0f