rename android_errorMessage to errorMessageAndroid
Following same implementation used with underlineColorAndroid
https://github.com/fabriziobertoglio1987/react-native/blob/15810e96d90e18dbd424666338fdec0127d403ed/Libraries/Components/TextInput/TextInput.js#L470
https://github.com/fabriziobertoglio1987/react-native/blob/15810e96d90e18dbd424666338fdec0127d403ed/Libraries/Components/TextInput/AndroidTextInputNativeComponent.js#L203

While with other components (for example prop android_ripple), we use
the android_ prefix to denote platform specific props.
https://github.com/facebook/react-native/pull/29157#issuecomment-645724520
https://github.com/fabriziobertoglio1987/react-native/blob/15810e96d90e18dbd424666338fdec0127d403ed/Libraries/Components/Pressable/Pressable.js#L177

In the case of TextInput we already have Platform Logic that detects
Android/iOS platform.
https://github.com/fabriziobertoglio1987/react-native/blob/15810e96d90e18dbd424666338fdec0127d403ed/Libraries/Components/TextInput/TextInput.js#L1268

For this reason TextInput component does not use android_ props and
instead uses this naming convention underlineColorAndroid.

To be noted that the prop underlineColorAndroid is passed to both iOS
and Android version, while other props have platform specific logic for
android and iOS.
https://github.com/fabriziobertoglio1987/react-native/blob/15810e96d90e18dbd424666338fdec0127d403ed/Libraries/Components/TextInput/TextInput.js#L1334
https://github.com/fabriziobertoglio1987/react-native/blob/15810e96d90e18dbd424666338fdec0127d403ed/Libraries/Components/TextInput/TextInput.js#L1293

Example of a prop that have a specific value on Android and is different
from iOS
https://github.com/fabriziobertoglio1987/react-native/blob/15810e96d90e18dbd424666338fdec0127d403ed/Libraries/Components/TextInput/TextInput.js#L1271

I decided to follow the same solution used in underlineColorAndroid.