Add missing `hermes-executor-debug` from the bundled version of React Native
Summary:
I've just realized that once we specified separate CMake target `hermes-executor-debug` and `hermes-executor-release` for the two RN build types, we actually missed that RN is distributed as a single .aar (the release one). So we're essentially skipping to distribute the -debug version of Hermes executor. This diff fixes it.

Changelog:
[Internal] [Changed] - Add missing `hermes-executor-debug` from the bundled version of React Native

Differential Revision: D35286345

fbshipit-source-id: 02cba47b3eb8d5f70bd0f00f71163889dd2c6b67