using std::optional<std::string>
rawProps collection has null value for the key. This is the special case which means that the prop was removed from the particular component instance and we have to reset it to some default value (which is not the same as sourceValue). Now the default value of the defaultValue (sic!) argument is a default value of the type of the value (which may be different from logical default value).

https://github.com/fabriziobertoglio1987/react-native-notes/issues/12#issuecomment-1084364074
https://github.com/fabriziobertoglio1987/react-native-notes/issues/12#issuecomment-1084316443