use value_or instead of pointer *paragraphAttributes.errorMessageAndroid
>A pointer however, is a variable that stores the memory address as its value.
>A pointer variable points to a data type (like int or string) of the same type, and is created with the * operator. The address of the variable you're working with is assigned to the pointer:
https://www.w3schools.com/cpp/cpp_pointers.asp

The solution was copied by allowFontScaling
```c++
  if (textAttributes.allowFontScaling.has_value()) {
    _textAttributes("allowFontScaling", *textAttributes.allowFontScaling);
  }
```
https://github.com/fabriziobertoglio1987/react-native/blob/53fe1c0d12b7d711819d3b79aac6eb99568cf86d/ReactCommon/react/renderer/attributedstring/conversions.h#L868

to solve the fat that folly::object() takes two parameters key and
value. The value can not be of type std::optional<std::string>

https://github.com/facebook/folly/blob/main/folly/docs/Dynamic.md

conversion to string with
folly::to<std::string>(paragraphAttributes.errorMessageAndroid) does not
work.

This solution first checks if paragraphAttributes.errorMessageAndroid is
not null with method has_value and then used value_or("").