Make Platform.select type simpler and more Flow-friendly
This way, if you write something like

    const x = Platform.select({ default: true, ios: false });

then Flow is able to successfully infer the type of `x` as `boolean`.
Before this change, it would infer `empty`, leading to loss of type
coverage.

The idea is that instead of trying to name the types of all the
individual options -- some of which may be absent, which then defeats
type inference -- we name just one type, the return type, and just
tell Flow that each option needs to be a subtype of that.

Note that this doesn't introduce any requirement that the types of
the different options be "the same", or any other new requirement on
callers.  Any call site was previously passing a valid
`PlatformSelectSpec<D, N, I>` for some types `D`, `N`, and `I`, and
by the new definition that will always be a valid
`PlatformSelectSpec<D | N | I>`.