Bump React Native Gradle plugin to 0.0.6
Summary:
We currently have some code on the RN Gradle Plugin that we need to ship.
There are both bugfixes needed for RN 0.68.1 and for the current nightly
(therefore needed for RN 0.69).

I've verified that this works on a fresh RN 0.68.0 install with `npm pack`
and triggering a build on a fresh setup from app template, with newArchEnabled set to true.

Changelog:
[Android] [Changed] - Bump React Native Gradle plugin to 0.0.6

Differential Revision: D35439444

fbshipit-source-id: 4d475c7d3b486c53c6bb290c0d16767faa7b566e