Pass node executable to codegen
Summary:
This pr adds a parameter to the `create-artifacts.js` script to accept the path to a node executable, falling back to `node` in case the parameter has not been passed.
Then, it passes the NODE_BINARY to the script in the `script_phases.sh` script.

This PR decouples the `node` environment from the system one and fixes a build issue in the new architecture when the environment has no `node`

Differential Revision: D35748497

fbshipit-source-id: 201daed0b31e4230f623631c634b5e2cd1c0aab6