This change helps in building React native with Hermes from source on Windows boxes.
Speicifically, this change adds
1. Explicitly use ninja with CMake to build the hermes compiler. VS generators works, but as VS is multiconfig, it requires the build flavor to be explicitly specified.
And the the build flavor of the hermes compiler and runtime builds should match which will require us to build multiple flavors of compiler for debug and ship. And it is tricky to map the compiler and runtime builds to match flavor.
Using Ninja generator avoids the complication.
2. CMake shipped with Android SDK can't find the python installation on Windows with FindPython, but it works fine with the older FindPythonInterp. The switch was made very recently in hermes repo (https://github.com/facebook/hermes/commit/5244f819b2f3949ca94a3a1bf75d54a8ed59d94a)
This change adds a hook for the developer to hint FindPython module to find Python interpreter on Windows